HeboEnums: Remove class decoration from enum type

+++ src/mqtt/message_stream_model.cpp
@@ -31,7 +31,7 @@ QVariant MessageStreamModel::data(const QModelIndex& index, int role) const {
   const auto& msg = this->messages_.at(index.row());
   switch (role) {
     case kTopicRole: { return msg.topic; }
-    case kQoSRole: { return static_cast<int>(msg.qos); }
+    case kQoSRole: { return msg.qos; }
     case kIsPublishRole: { return msg.is_publish; }
     case kTimestampRole: { return msg.timestamp; }
     case kPayloadRole: { return msg.payload; }
